Add event ID from `X-GitHub-Delivery` if present
If we see a `X-GitHub-Delivery` header use its value as the Event ID,
which allows us to use the GitHub Webhook page listing of recent events
to check against what the gosmee client has logged.

Further, we change the code to consolidate header handling so that each
header is handled once.