Optional fix and formatting for replay template
An optional fix to a comment which now properly references 8080, along
with a set of formatting fixes to "improve" the readability.

It also adds the use of the `-sSi` switches to the `curl` command so
that once can see the response headers (`-i) without the progress
meter (`-s`) while still seeing error messages (`-S`).