Add log entries with req headers to aid debugging
The new published log entries in both the client and server contain the
the original payload headers, and for the server the request ID,
allowing one to track down behaviors from know sources (e.g. GitHub
posts a list of events it sent to a web hook with the headers used).