Marshal time stamp as string
The server now encodes the time stamp at which it received the POST
operation as a string instead of using the default `float64` type
imposed by the standard `json` interface.

This also fixes the time stamp handling to correct sub-second values.

We also take the opportunity to record the time stamp as early as
possible and coalesce the code a bit.