feat(client): remove port from x-forwarded-for header
The X-Forwarded-For header is added to the outgoing request as
expected, but it includes the port number. This is incorrect
according to the specification and many endpoints reject any invalid
x-forwarded-for header in the request with "400 bad request". This
commit fixes this issue by removing the port number from the header.

Fixes https://github.com/chmouel/gosmee/issues/135